bug fixes, slice encode version

+++ src/mixer/utils/slice/cleaning/QuickCentroids.java
@@ -45,7 +45,7 @@ public class QuickCentroids {
     private final Random generator = new Random(0);
     private final AtomicInteger numActualClusters = new AtomicInteger(0);
     private float[][] centroids = null;
-    private float[] weights = null;
+    private int[] weights = null;
 
     public QuickCentroids(float[][] matrix, int numCentroids, long seed) {
         this.matrix = matrix;
@@ -118,7 +118,7 @@ public class QuickCentroids {
         }
         AtomicInteger currRowIndex = new AtomicInteger(0);
         centroids = new float[actualClusters.size()][matrix[0].length];
-        weights = new float[actualClusters.size()];
+        weights = new int[actualClusters.size()];
         ExecutorService executor = Executors.newFixedThreadPool(numCPUThreads);
         for (int l = 0; l < numCPUThreads; l++) {
             executor.execute(() -> {
@@ -166,7 +166,7 @@ public class QuickCentroids {
         return true;
     }
 
-    public float[] getWeights() {
+    public int[] getWeights() {
         return weights;
     }
 }
